+++ genvm/templates/multisig/types.go
@@ -4,15 +4,18 @@ import "github.com/spacemeshos/go-spacemesh/genvm/core"
 
 //go:generate scalegen
 
-// SpawnArguments ...
+// SpawnArguments contains a collection with PublicKeys.
 type SpawnArguments struct {
 	PublicKeys []core.PublicKey `scale:"type=StructArray"`
 }
 
+// Signatures is a collections of parts that must satisfy multisig
+// threshold requirement.
 type Signatures []Part
 
+// Part contains a reference to public key and signature from private key counterpart.
 type Part struct {
-	Ref uint64
+	Ref uint8
 	Sig core.Signature
 }
